Refactor custom config methods to new component
