+++ core/src/main/scala/org/apache/spark/sql/crossdata/XDSession.scala
@@ -217,13 +217,10 @@ object XDSession {
   /**
     * Builder for [[XDSession]].
     */
-  class Builder extends SparkSessionBuilder {
+  class Builder extends SparkSessionBuilder with BuilderEnhancer {
 
     private[this] val options = new scala.collection.mutable.HashMap[String, String]
     private[this] var userSuppliedContext: Option[SparkContext] = None
-    private[this] val ParentConfPrefix = "crossdata-core"
-    private[this] val SparkConfPrefix = "spark"
-    private[this] val CatalogConfPrefix = "catalog"
 
     override def config(key: String, value: String): Builder = synchronized {
       options += key -> value
@@ -233,24 +230,6 @@ object XDSession {
     override def config(key: String, value: Long): Builder = config(key, value)
     override def config(key: String, value: Double): Builder = config(key, value)
     override def config(key: String, value: Boolean): Builder = config(key, value)
-
-    def config(conf: Config): Builder = synchronized {
-      setSparkConf(conf)
-      setCatalogConf(conf)
-      this
-    }
-
-    def config(configFile: File): Builder = synchronized {
-      if (configFile.exists && configFile.canRead) {
-        log.info(s"Configuration file loaded ( ${configFile.getAbsolutePath} ).")
-        val conf = ConfigFactory.parseFile(configFile)
-        config(conf)
-      } else {
-        log.warn(s"Configuration file ( ${configFile.getAbsolutePath} ) is not accessible.")
-      }
-
-      this
-    }
 
     override def config(conf: SparkConf): Builder = synchronized {
       conf.getAll.foreach { case (k, v) => options += k -> v }
@@ -305,74 +284,11 @@ object XDSession {
         sc
       }
 
-      val catalogConf = extractCatalogConf()
+      val catalogConf = extractCatalogConf(options)
       val session = new XDSession(sparkContext, catalogConf)
       options.foreach { case (k, v) => session.sessionState.conf.setConfString(k, v) }
 
       session
-    }
-
-    /**
-      * Set Spark related configuration from Typesafe Config
-      * @param conf
-      */
-    private def setSparkConf(conf: Config): Unit = {
-      if (conf.hasPath(s"$ParentConfPrefix.$SparkConfPrefix")) {
-
-        val sparkConf: Config = conf
-          .getConfig(ParentConfPrefix)
-          .withOnlyPath(SparkConfPrefix)
-
-        import scala.collection.JavaConversions._
-
-        sparkConf
-          .entrySet()
-          .map(entry => (entry.getKey, entry.getValue.unwrapped().toString))
-          .foreach {
-            case (key, value) => config(key, value)
-          }
-
-      } else {
-        log.info(s"No spark configuration was found in configuration")
-      }
-    }
-
-    /**
-      * Set  Catalog configuration from Typesafe Config
-      * @param conf
-      */
-    private def setCatalogConf(conf: Config) = {
-      if (conf.hasPath(s"$ParentConfPrefix.$CatalogConfPrefix")) {
-        import scala.collection.JavaConversions._
-        conf
-          .withOnlyPath(s"$ParentConfPrefix.$CatalogConfPrefix")
-          .entrySet()
-          .map(entry => (entry.getKey, entry.getValue.unwrapped().toString))
-          .foreach {
-            case (key, value) => config(key, value)
-          }
-      } else {
-        log.info(s"No catalog configuration was found in configuration")
-      }
-    }
-
-    /**
-      * Extract Catalog configuration from options map
-      * @return Catalog configuration
-      */
-    private def extractCatalogConf(): Config = {
-      val catalogConf = options.filter {
-        case (key, _) => key.startsWith(s"$ParentConfPrefix.$CatalogConfPrefix")
-      }
-
-      import scala.collection.JavaConversions._
-      ConfigFactory.parseMap {
-        catalogConf
-          .map { t =>
-            (t._1.replaceFirst(s"$ParentConfPrefix.$CatalogConfPrefix.", ""), t._2)
-          }
-          .toMap[String, String]
-      }
     }
 
   }
